mend

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -81,7 +81,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public List<Film> getPopularFilms(int count) {
         return filmHashMap.values().stream()
-                .sorted(Comparator.comparingInt(film -> film.getUserIdLikes().size()))
+                .sorted(Comparator.comparingLong(film -> film.getUserIdLikes().size()))
                 .limit(count)
                 .toList();
     }
